FB-15a: extract persistApiStopReason and handleMistakeLimitReached into helpers
